requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
munch==4.0.0
pillow==10.4.0
requests==2.32.3
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
